changed questions endpoints naming

+++ src/main/java/com/jkh98/qna/controller/QuestionController.java
@@ -2,7 +2,9 @@ package com.jkh98.qna.controller;
 
 import com.jkh98.qna.exception.ResourceNotFoundException;
 import com.jkh98.qna.model.Question;
+import com.jkh98.qna.repository.CategoryRepository;
 import com.jkh98.qna.repository.QuestionRepository;
+import com.jkh98.qna.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,6 +22,12 @@ public class QuestionController {
     @Autowired
     private QuestionRepository questionRepository;
 
+    @Autowired
+    private CategoryRepository categoryRepository;
+
+    @Autowired
+    private UserRepository userRepository;
+
     @GetMapping("/questions")
     public Page<Question> getQuestions(Pageable pageable) {
         return questionRepository.findAll(pageable);
@@ -34,14 +42,42 @@ public class QuestionController {
         return questionMaybe.get();
     }
 
-    @PostMapping("/questions")
-    public Question createQuestion(@Valid @RequestBody Question question) {
-        return questionRepository.save(question);
+    @GetMapping("/categories/{categoryId}/questions")
+    public Page<Question> getQuestionsByCategory(@PathVariable UUID categoryId, Pageable pageable) {
+        if (!categoryRepository.existsById(categoryId)) {
+            throw new ResourceNotFoundException("Category not found with id " + categoryId);
+        }
+
+        return questionRepository.findByCategoryId(categoryId, pageable);
     }
 
-    @PutMapping("/questions/{questionId}")
-    public Question updateQuestion(@PathVariable UUID questionId,
+    @GetMapping("/users/{userId}/questions")
+    public Page<Question> getQuestionsByUser(@PathVariable UUID userId, Pageable pageable) {
+        if (!userRepository.existsById(userId)) {
+            throw new ResourceNotFoundException("User not found with id " + userId);
+        }
+
+        return questionRepository.findByUserId(userId, pageable);
+    }
+
+    @PostMapping("/categories/{categoryId}/questions")
+    public Question createQuestion(@PathVariable UUID categoryId, @Valid @RequestBody Question question) {
+        return categoryRepository.findById(categoryId)
+                .map(category -> {
+                    question.setCategory(category);
+                    return questionRepository.save(question);
+                }).orElseThrow(() -> new ResourceNotFoundException("Category not found with id " + categoryId));
+
+    }
+
+    @PutMapping("/categories/{categoryId}/questions/{questionId}")
+    public Question updateQuestion(@PathVariable UUID categoryId,
+                                   @PathVariable UUID questionId,
                                    @Valid @RequestBody Question questionRequest) {
+        if (!categoryRepository.existsById(categoryId)) {
+            throw new ResourceNotFoundException("Category not found with id " + questionId);
+        }
+
         return questionRepository.findById(questionId)
                 .map(question -> {
                     question.setTitle(questionRequest.getTitle());
@@ -50,8 +86,14 @@ public class QuestionController {
                 }).orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + questionId));
     }
 
-    @DeleteMapping("/questions/{id}")
-    public ResponseEntity<?> deleteQuestion(@PathVariable UUID questionId) {
+    @DeleteMapping("/categories/{categoryId}/questions/{questionId}")
+    public ResponseEntity<?> deleteQuestion(
+            @PathVariable UUID categoryId,
+            @PathVariable UUID questionId) {
+        if (!categoryRepository.existsById(categoryId)) {
+            throw new ResourceNotFoundException("Category not found with id " + questionId);
+        }
+
         return questionRepository.findById(questionId)
                 .map(question -> {
                     questionRepository.delete(question);
